[Update] change hibernate example code to single file; debug for review
