+++ plain-java-hibernate/src/main/java/com/pingcap/HibernateExample.java
@@ -1,0 +1,233 @@
+// Copyright 2022 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.pingcap;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import org.hibernate.JDBCException;
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
+import org.hibernate.cfg.Configuration;
+import org.hibernate.query.NativeQuery;
+import org.hibernate.query.Query;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.function.Function;
+
+@Entity
+@Table(name = "player_hibernate")
+class PlayerBean {
+    @Id
+    private String id;
+    @Column(name = "coins")
+    private Integer coins;
+    @Column(name = "goods")
+    private Integer goods;
+
+    public PlayerBean() {
+    }
+
+    public PlayerBean(String id, Integer coins, Integer goods) {
+        this.id = id;
+        this.coins = coins;
+        this.goods = goods;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public Integer getCoins() {
+        return coins;
+    }
+
+    public void setCoins(Integer coins) {
+        this.coins = coins;
+    }
+
+    public Integer getGoods() {
+        return goods;
+    }
+
+    public void setGoods(Integer goods) {
+        this.goods = goods;
+    }
+}
+
+/**
+ * Main class for the basic Hibernate example.
+ **/
+public class HibernateExample
+{
+    public static class PlayerDAO {
+        public static class NotEnoughException extends RuntimeException {
+            public NotEnoughException(String message) {
+                super(message);
+            }
+        }
+
+        // Run SQL code in a way that automatically handles the
+        // transaction retry logic so we don't have to duplicate it in
+        // various places.
+        public Object runTransaction(Session session, Function<Session, Object> fn) {
+            Object resultObject = null;
+
+            Transaction txn = session.beginTransaction();
+            try {
+                resultObject = fn.apply(session);
+                txn.commit();
+                System.out.println("APP: COMMIT;");
+            } catch (JDBCException e) {
+                System.out.println("APP: ROLLBACK BY JDBC ERROR;");
+                txn.rollback();
+            } catch (NotEnoughException e) {
+                System.out.printf("APP: ROLLBACK BY LOGIC; %s", e.getMessage());
+                txn.rollback();
+            }
+            return resultObject;
+        }
+
+        public Function<Session, Object> createPlayers(List<PlayerBean> players) throws JDBCException {
+            return session -> {
+                Integer addedPlayerAmount = 0;
+                for (PlayerBean player: players) {
+                    session.persist(player);
+                    addedPlayerAmount ++;
+                }
+                System.out.printf("APP: createPlayers() --> %d\n", addedPlayerAmount);
+                return addedPlayerAmount;
+            };
+        }
+
+        public Function<Session, Object> buyGoods(String sellId, String buyId, Integer amount, Integer price) throws JDBCException {
+            return session -> {
+                PlayerBean sellPlayer = session.get(PlayerBean.class, sellId);
+                PlayerBean buyPlayer = session.get(PlayerBean.class, buyId);
+
+                if (buyPlayer == null || sellPlayer == null) {
+                    throw new NotEnoughException("sell or buy player not exist");
+                }
+
+                if (buyPlayer.getCoins() < price || sellPlayer.getGoods() < amount) {
+                    throw new NotEnoughException("coins or goods not enough, rollback");
+                }
+
+                buyPlayer.setGoods(buyPlayer.getGoods() + amount);
+                buyPlayer.setCoins(buyPlayer.getCoins() - price);
+                session.persist(buyPlayer);
+
+                sellPlayer.setGoods(sellPlayer.getGoods() - amount);
+                sellPlayer.setCoins(sellPlayer.getCoins() + price);
+                session.persist(sellPlayer);
+
+                System.out.printf("APP: buyGoods --> sell: %s, buy: %s, amount: %d, price: %d\n", sellId, buyId, amount, price);
+                return 0;
+            };
+        }
+
+        public Function<Session, Object> getPlayerByID(String id) throws JDBCException {
+            return session -> session.get(PlayerBean.class, id);
+        }
+
+        public Function<Session, Object> printPlayers(Integer limit) throws JDBCException {
+            return session -> {
+                NativeQuery<PlayerBean> limitQuery = session.createNativeQuery("SELECT * FROM player_hibernate LIMIT :limit", PlayerBean.class);
+                limitQuery.setParameter("limit", limit);
+                List<PlayerBean> players = limitQuery.getResultList();
+
+                for (PlayerBean player: players) {
+                    System.out.println("\n[printPlayers]:");
+                    System.out.printf("    %-8s => %10s\n", "id", player.getId());
+                    System.out.printf("    %-8s => %d\n", "coins", player.getCoins());
+                    System.out.printf("    %-8s => %d\n", "goods", player.getGoods());
+                }
+                return 0;
+            };
+        }
+
+        public Function<Session, Object> countPlayers() throws JDBCException {
+            return session -> {
+                Query<Long> countQuery = session.createQuery("SELECT count(player_hibernate) FROM PlayerBean player_hibernate", Long.class);
+                return countQuery.getSingleResult();
+            };
+        }
+    }
+
+    public static void main(String[] args) {
+        // 1. Create a SessionFactory based on our hibernate.cfg.xml configuration
+        // file, which defines how to connect to the database.
+        SessionFactory sessionFactory
+                = new Configuration()
+                .configure("hibernate.cfg.xml")
+                .addAnnotatedClass(PlayerBean.class)
+                .buildSessionFactory();
+
+        try (Session session = sessionFactory.openSession()) {
+            // 2. And then, create DAO to manager your data
+            PlayerDAO playerDAO = new PlayerDAO();
+
+            // 3. Run some simple example
+
+            // Create a player
+            playerDAO.runTransaction(session, playerDAO.createPlayers(Collections.singletonList(
+                    new PlayerBean("test", 1, 1))));
+
+            // Get a player
+            PlayerBean testPlayer = (PlayerBean)playerDAO.runTransaction(session, playerDAO.getPlayerByID("test"));
+            System.out.printf("PlayerDAO.getPlayer:\n    => id: %s\n    => coins: %s\n    => goods: %s\n",
+                    testPlayer.getId(), testPlayer.getCoins(), testPlayer.getGoods());
+
+            // Count players amount
+            Long count = (Long)playerDAO.runTransaction(session, playerDAO.countPlayers());
+            System.out.printf("PlayerDAO.countPlayers:\n    => %d total players\n", count);
+
+            // Print 3 players
+            playerDAO.runTransaction(session, playerDAO.printPlayers(3));
+
+            // 4. Getting further
+            PlayerBean player1 = new PlayerBean("1", 100, 0);
+            PlayerBean player2 = new PlayerBean("2", 114514, 20);
+
+            // Insert two players "by hand", using INSERT on the backend.
+            int addedCount = (Integer)playerDAO.runTransaction(session,
+                    playerDAO.createPlayers(Arrays.asList(player1, player2)));
+            System.out.printf("PlayerDAO.createPlayers:\n    => %d total inserted players\n", addedCount);
+
+            // Player 1 want to buy 10 goods from player 2
+            // It will cost 500 coins, but player 1 can't afford it
+            System.out.println("\nPlayerDAO.buyGoods:\n    => this trade will fail");
+            Integer updatedCount = (Integer)playerDAO.runTransaction(session,
+                    playerDAO.buyGoods(player2.getId(), player1.getId(), 10, 500));
+            System.out.printf("PlayerDAO.buyGoods:\n    => %d total update players\n", updatedCount);
+
+            // So player 1 have to reduce his incoming quantity to two
+            System.out.println("\nPlayerDAO.buyGoods:\n    => this trade will success");
+            updatedCount = (Integer)playerDAO.runTransaction(session,
+                    playerDAO.buyGoods(player2.getId(), player1.getId(), 2, 100));
+            System.out.printf("PlayerDAO.buyGoods:\n    => %d total update players\n", updatedCount);
+        } finally {
+            sessionFactory.close();
+        }
+    }
+}
