--- plain-java-hibernate/src/main/java/com/pingcap/PlayerBean.java
@@ -1,62 +1,0 @@
-// Copyright 2022 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.pingcap;
-
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "player_hibernate")
-public class PlayerBean {
-    @Id
-    private String id;
-    @Column(name = "coins")
-    private Integer coins;
-    @Column(name = "goods")
-    private Integer goods;
-
-    public PlayerBean() {
-    }
-
-    public PlayerBean(String id, Integer coins, Integer goods) {
-        this.id = id;
-        this.coins = coins;
-        this.goods = goods;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public Integer getCoins() {
-        return coins;
-    }
-
-    public void setCoins(Integer coins) {
-        this.coins = coins;
-    }
-
-    public Integer getGoods() {
-        return goods;
-    }
-
-    public void setGoods(Integer goods) {
-        this.goods = goods;
-    }
-}
