+++ plain-java-jdbc/src/main/java/com/pingcap/JDBCExample.java
@@ -27,69 +27,6 @@ import java.util.*;
  **/
 public class JDBCExample
 {
-    public static void main(String[] args) {
-        // 1. Configure the example database connection.
-
-        // 1.1 Create a mysql data source instance
-        MysqlDataSource mysqlDataSource = new MysqlDataSource();
-
-        // 1.2 Set server name, port, database name , username and password
-        mysqlDataSource.setServerName("localhost");
-        mysqlDataSource.setPortNumber(4000);
-        mysqlDataSource.setDatabaseName("test");
-        mysqlDataSource.setUser("root");
-        mysqlDataSource.setPassword("");
-
-        // Or you can use jdbc string instead
-        // mysqlDataSource.setURL("jdbc:mysql://{host}:{port}/test?user={user}&password={password}");
-
-        // 2. And then, create DAO to manager your data
-        PlayerDAO dao = new PlayerDAO(mysqlDataSource);
-
-        // 3. Run some simple example
-
-        // Create a player, has a coin and a goods
-        dao.createPlayers(Collections.singletonList(new PlayerBean("test", 1, 1)));
-
-        // Get a player
-        PlayerBean testPlayer = dao.getPlayer("test");
-        System.out.printf("PlayerDAO.getPlayer:\n    => id: %s\n    => coins: %s\n    => goods: %s\n",
-                testPlayer.getId(), testPlayer.getCoins(), testPlayer.getGoods());
-
-        // Create players with bulk inserts, insert 1919 players totally, and per batch for 114 players
-        int addedCount = dao.bulkInsertRandomPlayers(1919, 114);
-        System.out.printf("PlayerDAO.bulkInsertRandomPlayers:\n    => %d total inserted players\n", addedCount);
-
-        // Count players amount
-        int count = dao.countPlayers();
-        System.out.printf("PlayerDAO.countPlayers:\n    => %d total players\n", count);
-
-        // Print 3 players
-        dao.printPlayers(3);
-
-        // 4. Getting further
-
-        // Player 1: id is "1", has only 100 coins
-        // Player 2: id is "2", has 114514 coins, and 20 goods
-        PlayerBean player1 = new PlayerBean("1", 100, 0);
-        PlayerBean player2 = new PlayerBean("2", 114514, 20);
-
-        // Insert two players "by hand", using INSERT on the backend.
-        addedCount = dao.createPlayers(Arrays.asList(player1, player2));
-        System.out.printf("PlayerDAO.createPlayers:\n    => %d total inserted players\n", addedCount);
-
-        // Player 1 want to buy 10 goods from player 2
-        // It will cost 500 coins, but player 1 can't afford it
-        System.out.println("\nPlayerDAO.buyGoods:\n    => this trade will fail");
-        int updatedCount = dao.buyGoods(player2.getId(), player1.getId(), 10, 500);
-        System.out.printf("PlayerDAO.buyGoods:\n    => %d total update players\n", updatedCount);
-
-        // So player 1 have to reduce his incoming quantity to two
-        System.out.println("\nPlayerDAO.buyGoods:\n    => this trade will success");
-        updatedCount = dao.buyGoods(player2.getId(), player1.getId(), 2, 100);
-        System.out.printf("PlayerDAO.buyGoods:\n    => %d total update players\n", updatedCount);
-    }
-
     public static class PlayerBean {
         private String id;
         private Integer coins;
@@ -428,5 +365,67 @@ public class JDBCExample
             return count;
         }
     }
-}
 
+    public static void main(String[] args) {
+        // 1. Configure the example database connection.
+
+        // 1.1 Create a mysql data source instance
+        MysqlDataSource mysqlDataSource = new MysqlDataSource();
+
+        // 1.2 Set server name, port, database name , username and password
+        mysqlDataSource.setServerName("localhost");
+        mysqlDataSource.setPortNumber(4000);
+        mysqlDataSource.setDatabaseName("test");
+        mysqlDataSource.setUser("root");
+        mysqlDataSource.setPassword("");
+
+        // Or you can use jdbc string instead
+        // mysqlDataSource.setURL("jdbc:mysql://{host}:{port}/test?user={user}&password={password}");
+
+        // 2. And then, create DAO to manager your data
+        PlayerDAO dao = new PlayerDAO(mysqlDataSource);
+
+        // 3. Run some simple example
+
+        // Create a player, has a coin and a goods
+        dao.createPlayers(Collections.singletonList(new PlayerBean("test", 1, 1)));
+
+        // Get a player
+        PlayerBean testPlayer = dao.getPlayer("test");
+        System.out.printf("PlayerDAO.getPlayer:\n    => id: %s\n    => coins: %s\n    => goods: %s\n",
+                testPlayer.getId(), testPlayer.getCoins(), testPlayer.getGoods());
+
+        // Create players with bulk inserts, insert 1919 players totally, and per batch for 114 players
+        int addedCount = dao.bulkInsertRandomPlayers(1919, 114);
+        System.out.printf("PlayerDAO.bulkInsertRandomPlayers:\n    => %d total inserted players\n", addedCount);
+
+        // Count players amount
+        int count = dao.countPlayers();
+        System.out.printf("PlayerDAO.countPlayers:\n    => %d total players\n", count);
+
+        // Print 3 players
+        dao.printPlayers(3);
+
+        // 4. Getting further
+
+        // Player 1: id is "1", has only 100 coins
+        // Player 2: id is "2", has 114514 coins, and 20 goods
+        PlayerBean player1 = new PlayerBean("1", 100, 0);
+        PlayerBean player2 = new PlayerBean("2", 114514, 20);
+
+        // Insert two players "by hand", using INSERT on the backend.
+        addedCount = dao.createPlayers(Arrays.asList(player1, player2));
+        System.out.printf("PlayerDAO.createPlayers:\n    => %d total inserted players\n", addedCount);
+
+        // Player 1 want to buy 10 goods from player 2
+        // It will cost 500 coins, but player 1 can't afford it
+        System.out.println("\nPlayerDAO.buyGoods:\n    => this trade will fail");
+        int updatedCount = dao.buyGoods(player2.getId(), player1.getId(), 10, 500);
+        System.out.printf("PlayerDAO.buyGoods:\n    => %d total update players\n", updatedCount);
+
+        // So player 1 have to reduce his incoming quantity to two
+        System.out.println("\nPlayerDAO.buyGoods:\n    => this trade will success");
+        updatedCount = dao.buyGoods(player2.getId(), player1.getId(), 2, 100);
+        System.out.printf("PlayerDAO.buyGoods:\n    => %d total update players\n", updatedCount);
+    }
+}
